LayoutTests/fast/dom/shadow/select-image-with-shadow-expected.txt

PASS selectedString is "KotoriPiyo"
PASS successfullyParsed is true

TEST COMPLETE
